fix: address review feedback on beacon mode PR
- Split beacon_helper.hpp into .hpp/.cpp with static library
- Use std::map for stable metadata key ordering
- Use int64_t for process_id parameter to avoid truncation
- Make entity_id a node parameter with default (reduces manifest coupling)
- Move import sys to top-level in launch file
- Increase beacon TTL from 10s to 15s (3x publish interval)
- Hardcode BEACON_MODE=none in CI service for deterministic tests
