+++ demos/sensor_diagnostics/src/beacon_helper.hpp
@@ -11,14 +11,11 @@
 #pragma once
 
 #include <chrono>
+#include <map>
 #include <memory>
 #include <string>
-#include <unordered_map>
 #include <vector>
 
-#include <unistd.h>
-
-#include "diagnostic_msgs/msg/key_value.hpp"
 #include "rclcpp/rclcpp.hpp"
 #include "ros2_medkit_msgs/msg/medkit_discovery_hint.hpp"
 
@@ -34,97 +31,15 @@ public:
     std::string display_name;
     std::string component_id;
     std::vector<std::string> function_ids;
-    std::unordered_map<std::string, std::string> metadata;
+    std::map<std::string, std::string> metadata;
   };
 
-  BeaconHelper(rclcpp::Node * node, const Config & config)
-  : node_(node), config_(config)
-  {
-    node_->declare_parameter("beacon_mode", "none");
-    mode_ = node_->get_parameter("beacon_mode").as_string();
-
-    // Cache process info (constant during lifetime)
-    pid_ = static_cast<uint32_t>(getpid());
-    char hostname_buf[256];
-    hostname_buf[sizeof(hostname_buf) - 1] = '\0';
-    if (gethostname(hostname_buf, sizeof(hostname_buf) - 1) == 0) {
-      hostname_ = hostname_buf;
-    }
-
-    if (mode_ == "topic") {
-      init_topic_beacon();
-    } else if (mode_ == "param") {
-      init_param_beacon();
-    } else if (mode_ != "none") {
-      RCLCPP_WARN(
-        node_->get_logger(),
-        "Unknown beacon_mode '%s', expected none/topic/param. Beacon disabled.",
-        mode_.c_str());
-    }
-  }
+  BeaconHelper(rclcpp::Node * node, const Config & config);
 
 private:
-  void init_topic_beacon()
-  {
-    beacon_pub_ = node_->create_publisher<ros2_medkit_msgs::msg::MedkitDiscoveryHint>(
-      "/ros2_medkit/discovery", 10);
-
-    beacon_timer_ = node_->create_wall_timer(
-      std::chrono::seconds(5),
-      [this]() { publish_beacon(); });
-
-    // Publish immediately on startup
-    publish_beacon();
-
-    RCLCPP_INFO(
-      node_->get_logger(), "Topic beacon enabled (entity: %s)",
-      config_.entity_id.c_str());
-  }
-
-  void init_param_beacon()
-  {
-    node_->declare_parameter("ros2_medkit.discovery.entity_id", config_.entity_id);
-    node_->declare_parameter("ros2_medkit.discovery.display_name", config_.display_name);
-    node_->declare_parameter("ros2_medkit.discovery.component_id", config_.component_id);
-    node_->declare_parameter("ros2_medkit.discovery.function_ids", config_.function_ids);
-    node_->declare_parameter(
-      "ros2_medkit.discovery.process_id", static_cast<int>(pid_));
-    node_->declare_parameter("ros2_medkit.discovery.process_name", node_->get_name());
-
-    if (!hostname_.empty()) {
-      node_->declare_parameter("ros2_medkit.discovery.hostname", hostname_);
-    }
-
-    for (const auto & [key, value] : config_.metadata) {
-      node_->declare_parameter("ros2_medkit.discovery.metadata." + key, value);
-    }
-
-    RCLCPP_INFO(
-      node_->get_logger(), "Parameter beacon enabled (entity: %s)",
-      config_.entity_id.c_str());
-  }
-
-  void publish_beacon()
-  {
-    auto msg = ros2_medkit_msgs::msg::MedkitDiscoveryHint();
-    msg.entity_id = config_.entity_id;
-    msg.display_name = config_.display_name;
-    msg.component_id = config_.component_id;
-    msg.function_ids = config_.function_ids;
-    msg.process_id = pid_;
-    msg.process_name = node_->get_name();
-    msg.hostname = hostname_;
-    msg.stamp = node_->now();
-
-    for (const auto & [key, value] : config_.metadata) {
-      diagnostic_msgs::msg::KeyValue kv;
-      kv.key = key;
-      kv.value = value;
-      msg.metadata.push_back(kv);
-    }
-
-    beacon_pub_->publish(msg);
-  }
+  void init_topic_beacon();
+  void init_param_beacon();
+  void publish_beacon();
 
   rclcpp::Node * node_;
   Config config_;
